Wrap many models
Wrap many models using a few simple solvers:

- Averaged: fit on each chunk, average the results. This works fine for
  most linear models, but isn't guaranteed to be correct (from what I
  understand).
- Chained: for estimators with a ``partial_fit`` method, fits by
  chaining calls to ``partial_fit`` together. Just provides a consistent
  interface to scikit-learn's existing functionality
- Gradient: just for the ``LogisticRegression`` for now, uses the
  previous gradient descent code.

Dask solvers are specified with a ``dask_solver`` keyword to the
estimator. Models are wrapped in a declarative way for now, which makes
this fairly easy to maintain. Wherever possible existing scikit-learn
functionality is used.

Solvers that can return ``Delayed`` objects will do so if
``compute=False`` is passed to the ``fit`` method. If this isn't
supported, a nice error is thrown.

Also removed the numpy compatible functionality, as it seemed
unnecessary, and complicated things.