Add a few methods to wrappers.
All methods take a `compute` kwarg, which is `True` by default. If
`False`, they return some type of delayed object. These also won't work
for all estimators, due to non-uniformities in scikit-learn estimators.
This indicates that it might be better to make (trivial) wrappers for
each of the estimators that these techniques (chaining and averaging)
make sense for, where the specifics of each one can be done correctly.

Methods added:
- predict
- decision_function
- score

Note that these only work/appear if the wrapped estimator has them.