Allow repeating cv extractions
Previously we performed the cv splits only once, and then reused them.
This is fast, but may result in extra memory usage, as each worker may
end up with multiple copies of the data in memory at once.

To fix this, we add the option to turn off the cacheing. This will take
more time (extracting a non-contiguous subset of the training data can
be expensive) but will only result in `n_threads + 1` copies of the data
on each worker, as opposed to `n_splits` copies from before. Cacheing is
turned on by default.

As an added benefit, the new structure plays better with data that's
already scattered throughout the cluster, as each call to fit directly
takes in the keys to the original (unsplit) dataset.